🛠️ README-Nextjs.md -> Edit page file updated 🔴 app/layout.tsx 🔴 app/page.tsx 🛠️ eslint.config.mjs -> Typescript config removed 🟢 jsconfig.json 🟢 next.config.mjs 🔴 next.config.ts 🛠️ package-lock.json -> Dependencies updated 🛠️ package.json -> Types and dev deps removed 🟢 src/app/layout.js 🟢 src/app/page.js 🔴 tsconfig.json 🔄 app/favicon.ico -> Renamed to src/app/ 🔄 app/globals.css -> Renamed to src/app/ 🔄 app/page.module.css -> Renamed to src/app/
