+++ src/app/layout.js
@@ -1,0 +1,25 @@
+import { Roboto } from "next/font/google";
+import "./globals.css";
+
+const RobotoSans = Roboto({
+  variable: "--font-roboto-sans",
+  subsets: ["latin"],
+  weight: ['300', '400', '500', '700'],
+  // weight: 'variable',
+  display: 'swap'
+});
+
+export const metadata = {
+  title: "Create Next App",
+  description: "Generated by create next app",
+};
+
+export default function RootLayout({ children }) {
+  return (
+    <html lang="en">
+      <body className={`${RobotoSans.variable}`}>
+        {children}
+      </body>
+    </html>
+  );
+}
